Fix id

+++ notifications_service/src/main/java/com/subOne/notifications_service/dto/notification/request/RequestUpdateNotificationsDto.java
@@ -6,5 +6,5 @@ import jakarta.validation.constraints.Size;
 import java.util.List;
 
 public record RequestUpdateNotificationsDto(
-        @NotNull(message = "ids not null") @Size(min = 1, max = 10, message = "Maximum of 10 notifications for updates and min of 1 notifications for updates") List<Long> ids) {
+        @NotNull(message = "ids not null") @Size(min = 1, max = 10, message = "Maximum of 10 notifications for updates and min of 1 notifications for updates") List<Integer> ids) {
 }
