+++ user_service/src/main/java/com/subOne/user_service/kafka/serviceProducer/KafkaService.java
@@ -7,6 +7,6 @@ import java.util.UUID;
 
 public interface KafkaService {
     Mono<Void> sendToTopic(String nameTopic, String value);
-    Mono<Void> sendToTopic(String nameTopic, Long value);
+    Mono<Void> sendToTopic(String nameTopic, Integer value);
     Mono<Void> sendToTopic(String nameTopic, UUID userId, SendNotificationDto kafkaDtoSendNotification);
 }
